working on new implementation of paging

+++ src/kernel/memory/memmap.h
@@ -7,7 +7,7 @@
 // print the memory map nicely
 void parse_multiboot_memmap(multiboot_info_t* mbinfo);
 // get the total amount of memory
-uint32_t get_total_mem(multiboot_info_t* mbinfo);
+uint64_t get_total_mem(multiboot_info_t* mbinfo);
 // get the number of memory entries
 uint32_t get_num_entries(multiboot_info_t* mbinfo);
 // set all type 1 regions as free
